Step 3: Finial Integration of request and obt data restructuring. (FIO #183)
