+++ contracts/fio.address/fio.address.cpp
@@ -1355,7 +1355,6 @@ namespace fioio {
         [[eosio::action]]
         void xferaddress(const string &fio_address, const string &new_owner_fio_public_key, const int64_t &max_fee,
                         const string &tpid, const name &actor) {
-            check(false, "Transfering a FIO address/domain is currently disabled");
             require_auth(actor);
             FioAddress fa;
             getFioAddressStruct(fio_address, fa);
